+++ src/lib/bluetooth/ace/src/bluetooth.cpp
@@ -565,7 +565,7 @@ int Bluetooth::attWriteCallback(hci_con_handle_t con_handle, uint16_t att_handle
                     memcpy(copy, payload.data(), payload.size());
                     // Adding a null terminator because it's expected downstream
                     copy[payload.size()] = 0;
-                    Bluetooth::instance->getBus().receive(GATAS::GatasConnectRx(pool, copy, payload.size() + 1));
+                    Bluetooth::instance->getBus().receive(GATAS::GatasConnectRx(pool, GATAS::GatasConnectTransport::Bluetooth, copy, payload.size() + 1));
                 } 
             }));
         // clang-format on